Fix add/remove items bug

+++ src/main/java/com/library/Library/service/CartService.java
@@ -9,6 +9,8 @@ import com.library.Library.repository.AppUserRepository;
 import com.library.Library.repository.BookRepository;
 import com.library.Library.repository.CartRepository;
 import lombok.AllArgsConstructor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -16,7 +18,6 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -25,23 +26,38 @@ public class CartService {
     private final CartRepository cartRepository;
     private final AppUserRepository appUserRepository;
     private final BookRepository bookRepository;
+    private final static Logger LOGGER = LoggerFactory
+            .getLogger(CartService.class);
 
     @Transactional
     public ResponseEntity<Cart> addItem(AddItemRequest request) {
         AppUser appUser = appUserRepository.findById(request.getAppUserId()).get();
         Book book = bookRepository.findBookById(request.getBookId()).get();
         List<Cart> cartList = cartRepository.findCartsByAppUser(appUser);
+        Boolean isBookExists = cartRepository.existsByBookAndAppUser(book, appUser);
         Cart cart;
         if (!cartList.isEmpty()) {
-            if (cartList.contains(book)) {
-                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "The book is already in cart!");
+            if (isBookExists) {
+                LOGGER.error("Here");
+                cart = cartRepository.findCartByBookAndAppUser(book, appUser);
+                cart.setQuantity(cart.getQuantity() + 1);
+            } else {
+                cart = new Cart(
+                        book,
+                        1,
+                        appUser,
+                        LocalDateTime.now()
+                );
+                return new ResponseEntity<>(cartRepository.save(cart), HttpStatus.OK);
             }
+        } else {
+            cart = new Cart(
+                    book,
+                    1,
+                    appUser,
+                    LocalDateTime.now()
+            );
         }
-        cart = new Cart(
-                book,
-                appUser,
-                LocalDateTime.now()
-        );
         return new ResponseEntity<>(cartRepository.save(cart), HttpStatus.OK);
     }
 
@@ -57,10 +73,14 @@ public class CartService {
         List<Cart> cartList = cartRepository.findCartsByAppUser(appUser);
         Book book = bookRepository.findBookById(request.getBookId()).get();
         if (!cartList.isEmpty()) {
-            Cart cart = cartRepository.findCartByBook(book);
-            if(cartList.contains(cart)){
-                cartRepository.deleteCartByBook(book);
-            } else throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "There is no " + book.getTitle() + "'s book in cart");
+            Cart cart = cartRepository.findCartByBookAndAppUser(book, appUser);
+            if (cartList.contains(cart)) {
+                cart.setQuantity(cart.getQuantity() - 1);
+                if (cart.getQuantity() == 0) {
+                    cartRepository.deleteCartByBook(book);
+                }
+            } else
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "There is no " + book.getTitle() + "'s book in cart");
         } else {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "There is no book found in cart");
         }
